pyproject.toml
[build-system]
requires = ["maturin>=1.9,<2.0"]
build-backend = "maturin"

[project]
name = "django-bolt"
version = "0.7.6"
description = "High-performance API framework for Django with Rust-powered endpoints delivering 60k+ RPS"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Farhan", email = "farhanalirazaazeemi@gmail.com" }
]
keywords = ["django", "api", "rust", "performance", "actix", "pyo3", "async"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    # "Programming Language :: Python :: 3.14", coming when msgspec has 3.14 support
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Rust",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "Django>=4.2",
    "click>=8.1",
    "httpx>=0.28.1",
    "msgspec>=0.20",
    "pyjwt>=2.10.1",
    "typing-extensions>=4.5.0",  # required openapi docs
    "pyyaml>=6.0", # required openapi docs
    "uvloop>=0.22.1; sys_platform != 'win32'"
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
nanodjango = ["nanodjango>=0.16"]

[project.urls]
Homepage = "https://github.com/FarhanAliRaza/django-bolt"
Repository = "https://github.com/FarhanAliRaza/django-bolt"
Documentation = "https://github.com/FarhanAliRaza/django-bolt#readme"
"Bug Tracker" = "https://github.com/FarhanAliRaza/django-bolt/issues"
[project.scripts]
django-bolt = "django_bolt.cli:main"

[project.entry-points."nanodjango"]
django-bolt = "django_bolt.nanodjango"

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "django_bolt._core"
python-source = "python"

[tool.uv.workspace]
members = []

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-django>=4.9.0",
    "pydantic>=2",
    "pyyaml>=5.2",
    "maturin>=1.9.6",
    "django==6.0.2",
    "ruff>=0.8",
    "vulture>=2.11", #for dead code detection
    "nanodjango>=0.16",
]
docs = [
    "zensical",
]

[tool.ruff]
src = ["python"]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S110",   # try-except-pass detected (no silent exception passing)
    "PLC0415", # import-outside-top-level (no imports inside functions/classes)
    "FIX002", # line contains TODO - must be resolved before merge
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for Depends)
    "ARG001", # unused function argument (common in handlers)
    "ARG002", # unused method argument
]

[tool.ruff.lint.isort]
known-first-party = ["django_bolt"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty.environment]
python-version = "3.12"
extra-paths = ["python"]
